tidy: rename subject folders to subjects
