+++ umich_catalog_indexing/lib/common/subjects/lc_subject.rb
