+++ umich_catalog_indexing/lib/common/subjects/non_lc_subject.rb
