+++ umich_catalog_indexing/lib/common/subjects/normalize.rb
