+++ umich_catalog_indexing/lib/common/subjects/subject.rb
@@ -1,0 +1,6 @@
+module Common
+  class Subjects
+    class Subject
+    end
+  end
+end
