+++ umich_catalog_indexing/spec/common/subjects/lc_subject_spec.rb
