+++ umich_catalog_indexing/spec/common/subjects/normalize_spec.rb
